refactor: share store contract validation helpers
